+++ src/AsyncHTTPSRequest_Debug_Generic.h
@@ -16,7 +16,7 @@
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
   You should have received a copy of the GNU General Public License along with this program.  If not, see <https://www.gnu.org/licenses/>.  
  
-  Version: 1.4.1
+  Version: 2.0.0
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -27,6 +27,7 @@
   1.3.0    K Hoang     23/01/2022 Enable compatibility with old code to include only AsyncHTTPSRequest_Generic.h
   1.4.0    K Hoang     11/02/2022 Add support to new ESP32-S3. Add LittleFS support to ESP32-C3. Use core LittleFS
   1.4.1    K Hoang     25/02/2022 Add example AsyncHTTPSRequest_ESP_Multi to demo connection to multiple addresses
+  2.0.0    K Hoang     27/02/2022 Breaking change to permit coexisting with AsyncHTTPRequest library. Add example to demo
  *****************************************************************************************************************************/
 
 #pragma once
@@ -35,9 +36,9 @@
 #define ASYNC_HTTPS_REQUEST_DEBUG_GENERIC_H
 
 #ifdef ASYNC_HTTPS_DEBUG_PORT
-  #define A_DBG_PORT      ASYNC_HTTPS_DEBUG_PORT
+  #define AHTTPS_DBG_PORT      ASYNC_HTTPS_DEBUG_PORT
 #else
-  #define A_DBG_PORT      Serial
+  #define AHTTPS_DBG_PORT      Serial
 #endif
 
 // Change _ASYNC_HTTPS_LOGLEVEL_ to set tracing and logging verbosity
@@ -56,10 +57,10 @@
 const char AHTTPS_MARK[] = "[AHTTPS] ";
 
 #define AHTTPS_PRINT_MARK   AHTTPS_PRINT(AHTTPS_MARK)
-#define AHTTPS_PRINT_SP     A_DBG_PORT.print(" ")
+#define AHTTPS_PRINT_SP     AHTTPS_DBG_PORT.print(" ")
 
-#define AHTTPS_PRINT        A_DBG_PORT.print
-#define AHTTPS_PRINTLN      A_DBG_PORT.println
+#define AHTTPS_PRINT        AHTTPS_DBG_PORT.print
+#define AHTTPS_PRINTLN      AHTTPS_DBG_PORT.println
 
 /////////////////////////////////////////////////////////